backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.12.1
email-validator==2.0.0 